[mono] XenAdmin: MainWindow: Disable adding HomePage tab content
For some reason, enabling that HomePage tab is the cause of connection lost right after the attempt to do connect to a server.
The HomePage tab shows nearly nothing useful than the add new server link. Actually the tab content is a web page that require a webbrowser library that doesn't look available at current days, and doesn't work on Mono, so just disable it. Significant negative impact to users on Mono platform doesn't expect.

Signed-off-by: Ilya Stolyarov <ila.embsys@gmail.com>